split github api adapters
