+++ new-github-indexer/domain/src/lib.rs
@@ -4,3 +4,6 @@ mod usecases;
 
 #[macro_use]
 extern crate async_trait;
+
+#[macro_use]
+extern crate serde;
