+++ new-github-indexer/infrastructure/src/adapters/github_api/adapters/mod.rs
@@ -1,0 +1,4 @@
+mod issue;
+mod pull_request;
+mod repo;
+mod user;
