+++ new-github-indexer/infrastructure/src/adapters/github_api/mod.rs
@@ -1,3 +1,29 @@
+mod adapters;
 mod error;
-mod github_api_adapter;
-pub use github_api_adapter::GithubApiAdapter;
+
+use std::fmt::Debug;
+
+use octocrab_indexer::{self, FromResponse, Octocrab};
+
+pub struct Client {
+	octocrab: Octocrab,
+}
+
+#[derive(Debug, Error)]
+pub enum Error {
+	#[error(transparent)]
+	Octocrab(#[from] octocrab_indexer::Error),
+}
+
+type Result<T> = std::result::Result<T, Error>;
+
+impl Client {
+	pub async fn get_as<U, R>(&self, route: U) -> Result<R>
+	where
+		U: AsRef<str> + Debug + Send,
+		R: FromResponse,
+	{
+		let response = self.octocrab.get(route, None::<&()>).await?;
+		Ok(response)
+	}
+}
