+++ new-github-indexer/infrastructure/src/adapters/mod.rs
@@ -1,4 +1,2 @@
-mod github_api;
-mod postgres_clean_storage;
-
-pub use github_api::GithubApiAdapter;
+pub mod github_api;
+pub mod postgres_clean_storage;
