+++ new-github-indexer/infrastructure/src/lib.rs
@@ -2,3 +2,6 @@ pub mod adapters;
 
 #[macro_use]
 extern crate async_trait;
+
+#[macro_use]
+extern crate thiserror;
